requirements.txt
admin-totals==1.0.1
adyengo==0.0.10
amqp==5.0.5
anyascii==0.1.7
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.1
Babel==2.9.0
beautifulsoup4==4.8.2
beautifultable==1.0.1
billiard==3.6.3.0
braintree==4.5.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
chromedriver-autoinstaller==0.2.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.17
django-3-jet==1.0.7
django-admin-autocomplete-filter==0.6.1
django-admin-filter==0.1
django-admin-notifications==0.6.3
django-admin-rangefilter==0.6.3
django-allauth==0.44.0
django-apscheduler==0.5.2
django-author==1.0.2
django-background-tasks==1.2.5
django-braces==1.13.0
django-celery==3.3.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-compat==1.0.15
django-cors-headers==3.5.0
django-countries==7.0
django-crispy-forms==1.10.0
django-crontab==0.7.1
django-csp==3.7
django-extensions==3.1.0
django-extra-views==0.11.0
django-filter==2.4.0
django-haystack==2.8.1
django-heroku==0.3.1
django-import-export==2.5.0
django-jet==1.0.8
django-js-asset==1.2.2
django-localflavor==3.0.1
django-model-utils==4.1.1
django-modelcluster==5.1
django-notifications-hq==1.6.0
django-nucleus==0.1.2
django-object-actions==3.0.1
django-oscar==2.0
django-oscar-api==2.0.2
django-oscar-paypal==2.0.0
django-oscar-promotions==1.0.0b1
django-oscar-support==0.1.0
django-oscar-wagtail==0.2.0
django-paypal==1.0.0
django-phonenumber-field==2.0.1
django-shortuuidfield==0.1.3
django-suit==2.0a1
django-tables2==1.21.2
django-taggit==1.3.0
django-timezone-field==4.1.1
django-treebeard==4.4
django-webline-notifications==1.0.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==5.8.0
filelock==3.0.12
geographiclib==1.50
geopy==2.1.0
gunicorn==20.0.4
h11==0.12.0
html5lib==1.1
httpcore==0.12.3
httpx==0.16.1
idna==2.10
ImageHash==4.2.0
ipaddress==1.0.23
jdatetime==3.6.2
jdcal==1.4.1
jsonfield==3.1.0
kombu==5.0.2
l18n==2020.6.1
MarkupPy==1.14
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
oscar-portation==0.1.1
persiantools==1.5.1
phonenumbers==8.12.17
Pillow==8.0.1
prompt-toolkit==3.0.14
psycopg2==2.8.6
psycopg2-binary==2.8.6
purl==1.5
pycparser==2.20
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-stdnum==1.15
python3-openid==3.2.0
pytube3==9.6.4
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
scipy==1.6.0
selenium==3.141.0
selenium-wire==2.1.2
setuptools-git==1.2
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
sorl-thumbnail==12.7.0
soupsieve==2.1
South==1.0.2
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.2
vine==5.0.0
virtualenv==20.3.1
wagtail==2.11.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
